Added Bitout_integral_value
Removed Bitset_output_iter and Bitset_forward_iter in favor of bitset_iter with is a random access iterator that can be used for both input and output.

Added Bitout_integral_value for writing the contents of a bitset into an integral value.

Added some comments to the library.

Significant clean up of whitespace style.

In test file, added comment about the need for unit tests (which don't rely on visual inspection of output).

Added "test" for Bitout_integral_value.